refactor: move and change AlgorithmSolver to algorithm-solver to match standard
