+++ packages/client/src/algorithm-solver/bstsolver.js
